AutoDoom: silenced all warnings from the game code (not the archiving libs) to have explicit casts from int64 to int32

+++ source/acs_func.cpp
@@ -564,7 +564,7 @@ int32_t ACS_GetThingVar(Mobj *thing, uint32_t var)
    case ACS_THINGVAR_MomY:           return thing->momy;
    case ACS_THINGVAR_MomZ:           return thing->momz;
    case ACS_THINGVAR_Pitch:          return thing->player ? thing->player->pitch >> 16 : 0;
-   case ACS_THINGVAR_PlayerNumber:   return thing->player ? thing->player - players : -1;
+   case ACS_THINGVAR_PlayerNumber:   return static_cast<int32_t>(thing->player ? thing->player - players : -1);
    case ACS_THINGVAR_SigilPieces:    return 0;
    case ACS_THINGVAR_TID:            return thing->tid;
    case ACS_THINGVAR_X:              return thing->x;
